pyproject.toml
[tool.poetry]
name = "demo-web-shop"
version = "0.1.0"
description = ""
authors = ["Daria Kudriavtseva <situation96@mail.ru>"]
readme = "README.md"
packages = [{include = "demo_web_shop"}]

[tool.pytest.ini_options]
log_cli = 1
log_level = "INFO"
log_cli_format = '%(asctime)s %(message)s'
log_cli_date_format = '%Y-%m-%d %H:%M:%S'
addopts = "-s -v --clean-alluredir --alluredir=allure-results --allure-no-capture"
markers = [
    "get: for test with GET request to API",
    "post: for test with POST request to API",
    "put: for test with PUT request to API",
    "patch: for test with PATCH request to API",
    "delete: for test with DELETE request to API"
]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.3.2"
selenium = "4.5.0"
selene = "2.0.0rc2"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
